// crates/sdk/bindings/assemblyscript/assembly/generated/UniScalarValue.ts
export enum UniScalarValueKind {

    Bool = 0,

    U8 = 1,

    I8 = 2,

    U16 = 3,

    I16 = 4,

    U32 = 5,

    I32 = 6,

    U64 = 7,

    U128 = 8,

    I64 = 9,

    I128 = 10,

    F32 = 11,

    F64 = 12,

    Char = 13,

    String = 14,

    Blob = 15,

    Numeric = 16,

    Date = 17,

    Time = 18,

    Timestamp = 19,

    TimestampTz = 20,

    Null = 21,

}

// AssemblyScript has no union types; the variant is a base class carrying the
// `kind` tag with one subclass per case. Codec functions take/return the base
// class and downcast with checked `as` casts.
export class UniScalarValue {
    kind: UniScalarValueKind = UniScalarValueKind.Bool;
}



export class UniScalarValueBool extends UniScalarValue {
    
    inner: bool = false;
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.Bool;
    }
}

export class UniScalarValueU8 extends UniScalarValue {
    
    inner: u8 = 0;
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.U8;
    }
}

export class UniScalarValueI8 extends UniScalarValue {
    
    inner: i8 = 0;
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.I8;
    }
}

export class UniScalarValueU16 extends UniScalarValue {
    
    inner: u16 = 0;
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.U16;
    }
}

export class UniScalarValueI16 extends UniScalarValue {
    
    inner: i16 = 0;
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.I16;
    }
}

export class UniScalarValueU32 extends UniScalarValue {
    
    inner: u32 = 0;
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.U32;
    }
}

export class UniScalarValueI32 extends UniScalarValue {
    
    inner: i32 = 0;
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.I32;
    }
}

export class UniScalarValueU64 extends UniScalarValue {
    
    inner: u64 = 0;
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.U64;
    }
}

export class UniScalarValueU128 extends UniScalarValue {
    
    inner: Uint8Array = new Uint8Array(0);
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.U128;
    }
}

export class UniScalarValueI64 extends UniScalarValue {
    
    inner: i64 = 0;
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.I64;
    }
}

export class UniScalarValueI128 extends UniScalarValue {
    
    inner: Uint8Array = new Uint8Array(0);
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.I128;
    }
}

export class UniScalarValueF32 extends UniScalarValue {
    
    inner: f32 = 0;
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.F32;
    }
}

export class UniScalarValueF64 extends UniScalarValue {
    
    inner: f64 = 0;
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.F64;
    }
}

export class UniScalarValueChar extends UniScalarValue {
    
    inner: string = "";
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.Char;
    }
}

export class UniScalarValueString extends UniScalarValue {
    
    inner: string = "";
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.String;
    }
}

export class UniScalarValueBlob extends UniScalarValue {
    
    inner: Uint8Array = new Uint8Array(0);
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.Blob;
    }
}

export class UniScalarValueNumeric extends UniScalarValue {
    
    inner: string = "";
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.Numeric;
    }
}

export class UniScalarValueDate extends UniScalarValue {
    
    inner: string = "";
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.Date;
    }
}

export class UniScalarValueTime extends UniScalarValue {
    
    inner: string = "";
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.Time;
    }
}

export class UniScalarValueTimestamp extends UniScalarValue {
    
    inner: string = "";
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.Timestamp;
    }
}

export class UniScalarValueTimestampTz extends UniScalarValue {
    
    inner: string = "";
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.TimestampTz;
    }
}

export class UniScalarValueNull extends UniScalarValue {
    
    constructor() {
        super();
        this.kind = UniScalarValueKind.Null;
    }
}